Luna: [BACKPORT] Implementation for Pixel 2 Launcher Merge
This add's support for Pixel 2 Launcher features. This is
a port of the backport done by @amirzaidi since he did the
initial backport, he get's authorship.

This is not all code from https://github.com/amirzaidi/Launcher3
just a pull for what we need and cleaned up a bit as an intial
indpendent launcher package. The backport was initially gradle
dependent, but i (@Sirhc95) have made several adjustments to add
support for the aosp build system (Makefile). This adds:

- Smartspace (Widget for Weather, Date, Clock, Events)
- New QSB in the hotseat
- Port of the Google Now Launcher Client
- Updated UI/UX for some areas such as AllAppsList and WidgetsList

We did not use the the new Settings menu implementation, instead features
are dumped in the aosp Settings menu.

Change-Id: I7ff69e2ab3e0d975304e1456f99b795b4bac4ce1