Luna: Implement Icon Pack & App Labeling interface
Adds support for changing application icons, as well as changing the
label of the application. You can change all of the application icons,
or choose which apps you want to change.

Editing the application label is simple. Long pressing the application
and tapping the pencil icon will bring up the editing menu to change the
application label. You can change the icon by pressing the current icon
and select the icon pack you wish to choose the icon from. This will
then bring up a icon chooser view to select any icon available from the
chosen icon pack.

ezio84:
- port to our Oreo Pixel Launcher mod
- fix blank icon for Oreo AdaptiveIconDrawable icons
- move settings to our fragments and shortcuts popups
- dinamycally theme icon edit dialogs and popups
- todo: when using a custom icon pack that doesn't have an icon for an
app with AdaptiveIconDrawable icon, the square shape is set for the
background drawable if the icon pack doesn't have an iconback drawable.
It'd be good instead to mask the background to the shape enabled by the
user. See IconsHandler.java getDefaultAppDrawable, we could mask
backgroundDr to:
AdaptiveIconDrawable backgroundDr = (AdaptiveIconDrawable)
   mContext.getDrawable(R.drawable.adaptive_icon_drawable_wrapper).mutate();
the same issue happens when resetting the icon if using the stock icon
pack, that's why atm i disabled icon customization for it
- todo2: cleanup and improve stuff

Sirhc95: Adapted for Luna Launcher

Change-Id: I712612d34cef512b492decb6c71689d44fb1a4a9